add auth

+++ server/app/controllers/ApiCartController.scala
@@ -1,38 +1,36 @@
 package controllers
 
 import javax.inject._
-import models.{CartProductRepository, ProductRepository, UserRepository}
+import models.{ CartProductRepository, ProductRepository, UserRepository }
 import play.api.libs.json._
 import play.api.mvc._
 import play.filters.csrf.CSRF
 
 import scala.concurrent.duration.Duration
-import scala.concurrent.{Await, ExecutionContext, Future}
-
+import scala.concurrent.{ Await, ExecutionContext, Future }
 
 @Singleton
-class ApiCartController @Inject()(cartProductRepository: CartProductRepository,
-                               userRepository: UserRepository,
-                               productRepository: ProductRepository,
-                               cc: MessagesControllerComponents) (implicit ec: ExecutionContext) extends MessagesAbstractController(cc) {
+class ApiCartController @Inject() (
+  cartProductRepository: CartProductRepository,
+  userRepository: UserRepository,
+  productRepository: ProductRepository,
+  cc: MessagesControllerComponents)(implicit ec: ExecutionContext) extends MessagesAbstractController(cc) {
 
   def getToken = Action { implicit request =>
     val token = CSRF.getToken.get.value
     Ok(token)
   }
 
-  def create:Action[AnyContent] = Action.async { implicit request =>
-    val params = request.queryString.map { case (k,v) => k -> v.mkString }
+  def create: Action[AnyContent] = Action.async { implicit request =>
+    val params = request.queryString.map { case (k, v) => k -> v.mkString }
     if (!params.contains("userId")) {
       Future(Ok("Error during adding product to bucket"))
-    }
-    else if (!params.contains("products")) {
+    } else if (!params.contains("products")) {
       Future(Ok("Error during adding product to bucket"))
-    }
-    else {
-      val prodArray = params("products").replaceAll(" ", "").split( ',' )
+    } else {
+      val prodArray = params("products").replaceAll(" ", "").split(',')
       val longProdArray = prodArray.map(_.toLong)
-      val userId = params("userId").toInt
+      val userId = params("userId")
 
       for (prodId <- longProdArray) {
         userRepository.exists(userId).map(userExists => {
@@ -49,24 +47,22 @@ class ApiCartController @Inject()(cartProductRepository: CartProductRepository,
 
   def read: Action[AnyContent] = Action.async { implicit request =>
     val carts = cartProductRepository.list()
-    carts.map( cart => Ok( Json.toJson(cart) ) )
+    carts.map(cart => Ok(Json.toJson(cart)))
   }
 
   def readById: Action[AnyContent] = Action { implicit request =>
 
-    val params = request.queryString.map { case (k,v) => k -> v.mkString }
+    val params = request.queryString.map { case (k, v) => k -> v.mkString }
     if (!params.contains("id")) {
       Ok("No id parameter in query")
-    }
-    else {
-      val id = params("id").toLong
+    } else {
+      val id = params("id")
       val cart = Await.result(cartProductRepository.getByUserId(id), Duration.Inf)
-      val products = cart.map( c => c._2 )
+      val products = cart.map(c => c._2)
 
       val res = JsObject(Seq(
         ("id", Json.toJson(id)),
-        ("products", Json.toJson(products))
-      ))
+        ("products", Json.toJson(products))))
 
       Ok(res)
     }
@@ -80,37 +76,32 @@ class ApiCartController @Inject()(cartProductRepository: CartProductRepository,
 
   // By user id
   def delete = Action { implicit request =>
-    val params = request.queryString.map { case (k,v) => k -> v.mkString }
+    val params = request.queryString.map { case (k, v) => k -> v.mkString }
 
     if (!params.contains("userId")) {
       Ok("No userId parameter in query")
-    }
-    else {
+    } else {
       try {
-        cartProductRepository.deleteUser(params("userId").toLong)
+        cartProductRepository.deleteUser(params("userId"))
         Ok("Bucket for user with id " + params("userId") + " deleted!")
-      }
-      catch {
+      } catch {
         case e: NumberFormatException => Ok("Id has to be integer")
       }
     }
   }
 
   def deleteProductForUser = Action { implicit request =>
-    val params = request.queryString.map { case (k,v) => k -> v.mkString }
+    val params = request.queryString.map { case (k, v) => k -> v.mkString }
 
     if (!params.contains("userId")) {
       Ok("Error during deleting product!")
-    }
-    else if (!params.contains("productId")) {
+    } else if (!params.contains("productId")) {
       Ok("Error during deleting product!")
-    }
-    else {
+    } else {
       try {
-        cartProductRepository.deleteProductForUser(params("productId").toLong, params("userId").toLong)
+        cartProductRepository.deleteProductForUser(params("productId").toLong, params("userId"))
         Ok("Product with id " + params("productId") + " deleted!")
-      }
-      catch {
+      } catch {
         case e: NumberFormatException => Ok("Error during deleting product!")
       }
     }
