+++ server/app/controllers/ApiOrderController.scala
@@ -1,19 +1,20 @@
 package controllers
 
 import javax.inject._
-import models.{Order, OrderProductRepository, OrderRepository, Product, ProductRepository}
+import models.{ Order, OrderProductRepository, OrderRepository, Product, ProductRepository }
 import play.api.libs.json._
 import play.api.mvc._
 import play.filters.csrf.CSRF
 
 import scala.concurrent.duration.Duration
-import scala.concurrent.{Await, ExecutionContext, Future}
+import scala.concurrent.{ Await, ExecutionContext, Future }
 
 @Singleton
-class ApiOrderController @Inject()(productRepository: ProductRepository,
-                                orderRepository: OrderRepository,
-                                orderProductRepository: OrderProductRepository,
-                                cc: MessagesControllerComponents)(implicit ec: ExecutionContext)
+class ApiOrderController @Inject() (
+  productRepository: ProductRepository,
+  orderRepository: OrderRepository,
+  orderProductRepository: OrderProductRepository,
+  cc: MessagesControllerComponents)(implicit ec: ExecutionContext)
   extends MessagesAbstractController(cc) {
 
   def getToken = Action { implicit request =>
@@ -21,24 +22,20 @@ class ApiOrderController @Inject()(productRepository: ProductRepository,
     Ok(token)
   }
 
-  def create:Action[AnyContent] = Action.async { implicit request =>
-    val params = request.queryString.map { case (k,v) => k -> v.mkString }
+  def create: Action[AnyContent] = Action.async { implicit request =>
+    val params = request.queryString.map { case (k, v) => k -> v.mkString }
     if (!params.contains("userId")) {
       Future(Ok("No userId parameter in query"))
-    }
-    else if (!params.contains("paymentId")) {
+    } else if (!params.contains("paymentId")) {
       Future(Ok("No paymentId parameter in query"))
-    }
-    else if (!params.contains("status")) {
+    } else if (!params.contains("status")) {
       Future(Ok("No status parameter in query"))
-    }
-    else if (!params.contains("products")) {
+    } else if (!params.contains("products")) {
       Future(Ok("No products parameter in query"))
-    }
-    else {
-      val prodArray = params("products").replaceAll(" ", "").split( ',' )
+    } else {
+      val prodArray = params("products").replaceAll(" ", "").split(',')
       val longProdArray = prodArray.map(_.toLong)
-      val prodId = orderRepository.create(params("userId").toLong, params("paymentId").toLong, params("status"))
+      val prodId = orderRepository.create(params("userId"), params("paymentId").toLong, params("status"))
       prodId.map(id => {
         for (prodId <- longProdArray) {
           productRepository.exists(prodId).map(exists => {
@@ -60,11 +57,10 @@ class ApiOrderController @Inject()(productRepository: ProductRepository,
 
   def readById: Action[AnyContent] = Action { implicit request =>
 
-    val params = request.queryString.map { case (k,v) => k -> v.mkString }
+    val params = request.queryString.map { case (k, v) => k -> v.mkString }
     if (!params.contains("id")) {
       Ok("No id parameter in query")
-    }
-    else {
+    } else {
 
       try {
         val orderProducts = Await.result(orderProductRepository.getByOrderId(params("id").toLong), Duration.Inf)
@@ -74,12 +70,10 @@ class ApiOrderController @Inject()(productRepository: ProductRepository,
 
         val res = JsObject(Seq(
           ("order", jsonOrders),
-          ("products", Json.toJson(products))
-        ))
+          ("products", Json.toJson(products))))
 
         Ok(res)
-      }
-      catch {
+      } catch {
         case e: NoSuchElementException => Ok("No element with such id")
       }
 
@@ -89,25 +83,22 @@ class ApiOrderController @Inject()(productRepository: ProductRepository,
 
   def readByUserId: Action[AnyContent] = Action { implicit request =>
 
-    val params = request.queryString.map { case (k,v) => k -> v.mkString }
+    val params = request.queryString.map { case (k, v) => k -> v.mkString }
     if (!params.contains("userId")) {
       Ok("No userId parameter in query")
-    }
-    else {
+    } else {
 
       try {
-        val orderProducts = Await.result(orderProductRepository.getByUserId(params("userId").toLong), Duration.Inf)
+        val orderProducts = Await.result(orderProductRepository.getByUserId(params("userId")), Duration.Inf)
 
-        val id = params("userId").toLong
-        val orders = orderProducts.map( c => c._1 )
+        val id = params("userId")
+        val orders = orderProducts.map(c => c._1)
         val res = JsObject(Seq(
-                      ("userId", Json.toJson(id)),
-                      ("orders", Json.toJson(orders))
-                    ))
+          ("userId", Json.toJson(id)),
+          ("orders", Json.toJson(orders))))
 
         Ok(res)
-      }
-      catch {
+      } catch {
         case e: NoSuchElementException => Ok("No element with such id")
       }
 
@@ -116,12 +107,11 @@ class ApiOrderController @Inject()(productRepository: ProductRepository,
   }
 
   def update = Action.async { implicit request =>
-    val params = request.queryString.map { case (k,v) => k -> v.mkString }
+    val params = request.queryString.map { case (k, v) => k -> v.mkString }
 
     if (!params.contains("id")) {
       Future(Ok("No id parameter in query"))
-    }
-    else {
+    } else {
 
       try {
         val id = params("id").toLong
@@ -129,12 +119,13 @@ class ApiOrderController @Inject()(productRepository: ProductRepository,
 
         orders.map(order => order match {
           case Some(o) => {
-            val userId = if (params.contains("userId")) params("userId").toLong else o.userId
+            val userId = if (params.contains("userId")) params("userId") else o.userId
             val paymentId = if (params.contains("paymentId")) params("paymentId").toLong else o.paymentId
             val status = if (params.contains("status")) params("status") else o.status
             val products = if (params.contains("products")) params("products").replaceAll(" ", "")
               .split(',')
-              .map(_.toLong) else Array[Long]()
+              .map(_.toLong)
+            else Array[Long]()
 
             if (products.size != 0) orderProductRepository.deleteOrder(id) // if products are not empty, delete old one
 
@@ -151,8 +142,7 @@ class ApiOrderController @Inject()(productRepository: ProductRepository,
           }
           case None => Ok("No object with such id")
         })
-      }
-      catch {
+      } catch {
         case e: NumberFormatException => Future(Ok("Id has to be integer"))
       }
 
@@ -160,18 +150,16 @@ class ApiOrderController @Inject()(productRepository: ProductRepository,
   }
 
   def delete = Action { implicit request =>
-    val params = request.queryString.map { case (k,v) => k -> v.mkString }
+    val params = request.queryString.map { case (k, v) => k -> v.mkString }
 
     if (!params.contains("id")) {
       Ok("No id parameter in query")
-    }
-    else {
+    } else {
       try {
         orderProductRepository.deleteOrder(params("id").toLong)
         orderRepository.delete(params("id").toLong)
         Ok("Order deleted!")
-      }
-      catch {
+      } catch {
         case e: NumberFormatException => Ok("Id has to be integer")
       }
     }
